+++ src/main/java/io/github/queerbric/inspecio/tooltip/ArmorStandTooltipComponent.java
@@ -36,7 +36,7 @@ import java.util.Optional;
  * Represents an armor stand tooltip. Displays an armor stand and its armor.
  *
  * @author Zailer43
- * @version 1.6.0
+ * @version 1.8.0
  * @since 1.1.0
  */
 public class ArmorStandTooltipComponent extends EntityTooltipComponent<InspecioConfig.EntityConfig> {
@@ -67,10 +67,10 @@ public class ArmorStandTooltipComponent extends EntityTooltipComponent<InspecioC
 	}
 
 	@Override
-	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer, int z) {
+	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer) {
 		if (this.shouldRender()) {
 			matrices.push();
-			matrices.translate(30, 0, z);
+			matrices.translate(30, 0, 0);
 			((EntityAccessor) this.entity).setTouchingWater(true);
 			this.entity.setVelocity(1.f, 1.f, 1.f);
 			this.renderEntity(matrices, x + 20, y + 12, this.entity, 0, this.config.shouldSpin(), true, 180.f);