+++ src/main/java/io/github/queerbric/inspecio/tooltip/ArmorTooltipComponent.java
@@ -49,7 +49,7 @@ public class ArmorTooltipComponent implements ConvertibleTooltipData, TooltipCom
 	}
 
 	@Override
-	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer, int z) {
+	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer) {
 		RenderSystem.setShaderTexture(0, InGameHud.GUI_ICONS_TEXTURE);
 		for (int i = 0; i < this.prot / 2; i++) {
 			DrawableHelper.drawTexture(matrices, x + i * 9, y, 34, 9, 9, 9, 256, 256);