+++ src/main/java/io/github/queerbric/inspecio/tooltip/BannerTooltipComponent.java
@@ -82,10 +82,10 @@ public class BannerTooltipComponent implements ConvertibleTooltipData, TooltipCo
 	}
 
 	@Override
-	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer, int z) {
+	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer) {
 		DiffuseLighting.setupFlatGuiLighting();
 		matrices.push();
-		matrices.translate(x + 8, y + 8, z);
+		matrices.translate(x + 8, y + 8, 0);
 		matrices.push();
 		matrices.translate(0.5, 16, 0);
 		matrices.scale(6, -6, 1);