+++ src/main/java/io/github/queerbric/inspecio/tooltip/FoodTooltipComponent.java
@@ -51,7 +51,7 @@ public record FoodTooltipComponent(FoodComponent component) implements Convertib
 	}
 
 	@Override
-	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer, int z) {
+	public void drawItems(TextRenderer textRenderer, int x, int y, MatrixStack matrices, ItemRenderer itemRenderer) {
 		var foodConfig = Inspecio.getConfig().getFoodConfig();
 
 		RenderSystem.setShaderTexture(0, InGameHud.GUI_ICONS_TEXTURE);